Move test utilities to Fixtures directory and remove SwiftMailer support
