+++ tests/Functional/DefaultSenderTest.php
@@ -13,8 +13,8 @@ declare(strict_types=1);
 
 namespace Sylius\Bundle\MailerBundle\Tests\Functional;
 
-use Sylius\Bundle\MailerBundle\Tests\Provider\MessagesProvider;
-use Sylius\Bundle\MailerBundle\Tests\Purger\SentMessagesPurger;
+use Sylius\Bundle\MailerBundle\Tests\Fixtures\MessagesProvider;
+use Sylius\Bundle\MailerBundle\Tests\Fixtures\SentMessagesPurger;
 use Sylius\Component\Mailer\Sender\SenderInterface;
 use Symfony\Bundle\FrameworkBundle\Test\KernelTestCase;
 use Symfony\Bundle\FrameworkBundle\Test\MailerAssertionsTrait;
